Se modifican funciones de subir y eliminar archivos para utilizarse con distintos módulos; se implementa dlg_borrar_archivos.

+++ application/controllers/Archivos.php
@@ -100,11 +100,14 @@ class Archivos extends CI_Controller {
     public function subir()
     {
         if ($this->session->userdata('logueado')) {
-            $ruta = $this->input->post('ruta');
+
+            $dir_docs = $this->input->post('dir_docs');
             $nombre_archivo = $this->input->post('nombre_archivo');
             $tipo_archivo = $this->input->post('tipo_archivo');
+            $url_actual = $this->input->post('url_actual');
+
             $config = array();
-            $config['upload_path'] = $ruta;
+            $config['upload_path'] = $dir_docs;
             $config['file_name'] = $nombre_archivo;
             $config['allowed_types'] = $tipo_archivo;
             $config['max_size'] = '10240';
@@ -139,42 +142,50 @@ class Archivos extends CI_Controller {
                 $this->bitacora_model->guardar($data);
 
             }
-            redirect('gestion');
+            redirect($url_actual);
         } else {
             redirect('inicio/login');
         }
     }
 
-    public function eliminar($nombre_archivo)
+    public function eliminar()
     {
         if ($this->session->userdata('logueado')) {
-            // registro en bitacora
-            $separador = ' -> ';
-            $usuario = $this->session->userdata('usuario');
-            $nom_usuario = $this->session->userdata('nom_usuario');
-            $nom_dependencia = $this->session->userdata('nom_dependencia');
-            $entidad = 'archivos';
-            $valor = $nombre_archivo;
-            $accion = 'eliminó';
-            $data = array(
-                'fecha' => date("Y-m-d"),
-                'hora' => date("H:i"),
-                'origen' => $_SERVER['REMOTE_ADDR'],
-                'usuario' => $usuario,
-                'nom_usuario' => $nom_usuario,
-                'nom_dependencia' => $nom_dependencia,
-                'accion' => $accion,
-                'entidad' => $entidad,
-                'valor' => $valor
-            );
-            $this->bitacora_model->guardar($data);
 
-            // Eliminar archivo
-            $nombre_archivo_fs = './doc/' . $nombre_archivo ;
-            $status = unlink($nombre_archivo_fs) ? 'Se eliminó el archivo '.$nombre_archivo_fs : 'Error al eliminar el archivo '.$nombre_archivo_fs;
-            echo $status;
+            $datos = $this->input->post();
+            if ($datos) {
+                $dir_docs = $datos['dir_docs'];
+                $nombre_archivo = $datos['nombre_archivo'];
+                $nombre_archivo_fs = './' . $dir_docs . $nombre_archivo ;
+                $url_actual = $datos['url_actual'];
 
-            redirect('gestion');
+                // registro en bitacora
+                $separador = ' -> ';
+                $usuario = $this->session->userdata('usuario');
+                $nom_usuario = $this->session->userdata('nom_usuario');
+                $nom_dependencia = $this->session->userdata('nom_dependencia');
+                $entidad = 'archivos';
+                $valor = $nombre_archivo;
+                $accion = 'eliminó';
+                $data = array(
+                    'fecha' => date("Y-m-d"),
+                    'hora' => date("H:i"),
+                    'origen' => $_SERVER['REMOTE_ADDR'],
+                    'usuario' => $usuario,
+                    'nom_usuario' => $nom_usuario,
+                    'nom_dependencia' => $nom_dependencia,
+                    'accion' => $accion,
+                    'entidad' => $entidad,
+                    'valor' => $valor
+                );
+                $this->bitacora_model->guardar($data);
+
+                // Eliminar archivo
+                $status = unlink($nombre_archivo_fs) ? 'Se eliminó el archivo '.$nombre_archivo_fs : 'Error al eliminar el archivo '.$nombre_archivo_fs;
+                echo $status;
+            }
+
+            redirect($url_actual);
         } else {
             redirect('inicio/login');
         }
